project/project/requirements.txt
Django==3.0.6
djangorestframework==3.7.3
googletrans==2.4.0